refactor: rename sort transactions util
